Promote the example interpreter to part of the codebase
This commit promotes the `ph7_interp.c` example to a more active
part of the codebase.

In order to keep source code organized, we moved the ph7 engine
to its own folder, and created a new `phl` folder in which the
interpreter will live separately.

Build files were changed accordingly, and our previous changes
to the example file were reverted.