Setup NT test dependencies (OpenCppCoverage and PHP)
This is the first step in being able to retrieve coverage from
MSVC builds, installing OpenCppCoverage.

Due to the use of #ifdefs, there is code only reachable if we
run them on the platform that those ifdefs target.

Small refactorings were made to the build-aux folder to enable
running OpenCppCoverage directly from WSL, and cosmetic changes.

CI was made to download OpenCppCoverage, but in this step we
are not going to use it to produce a report yet.

A script to install PHP was also added, although GitHub CI
already has PHP installed by default.