Add set_error_handler implementation
This commit makes the first steps into more elaborate error handling
for the ph7 engine by enabling the incomplete set_error_handler()
implementation and finishing its handler.

The original idea seems to be able to expose error handling to
the ph7 embedding API, however, this is not explored in this stage.