Improve windows build coverage output
The `cobertura_xml_to_lcov_info.php` file was relying on the
implied structure used by `dev.sh` (WSL wrapper), and thus, not
fully doing its job on pure windows builds (such as the CI),
leaving coverage files with absolute paths.

There was a hack in the GitHub workflow to fix this using sed,
but it was being applied in the wrong place, generating some
duplication in some reports.

This commit fixes the script and removes the hack, hopefully
improving the list of covered files to not feature duplicates.

It's likely that this won't affect final coverage metrics,
because lcov and genhtml were working around the duplication
anyways. However, this makes our builds cleaner, and some
artifacts (such as comments in build summaries) better.