Increase code coverage
Mostly constants. For flag constants, we don't care about their value,
just if it is present or not.

For math constants, we do a low precision check (first few digits)
to ensure the value is sound, but allow for precision differences
across PH7/PHL.

We are also updating tests to ensure all of them end with a blank
new line.