Re-enable --target-executable for phpt.php runner
This commit changes the popen function on WINNT to always use
cmd.exe, making it even closer to official PHP behavior.

We also re-enabled the `--target-executable` option for the
phpt.php runner, which allows us to run each test in a separate
proccess. This will be necessary to create tests that verify
behavior for fatal and syntax errors, among other unrecoverable
types of tests in the future.